requirements.txt
requests==2.31.0
python-dotenv==1.0.1
rich==13.7.0
Flask==3.0.2
flask-cors==4.0.0
